requirements.txt
# Python 3.10+ required
# Discord bot
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0

# Rocket League API
rlapi>=0.8.0b1

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Web UI (Phase 4)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
